feat(database): rename database
from
IDBWrapper-${databaseName}
to
IDBWrapper-${databaseName}_{development,production}
