+++ lib/utils/configs.js
@@ -1,4 +1,4 @@
-import { databaseName } from '../../package.json';
+import databaseName from '../utils/database-name-wrapper';
 import levelup from 'levelup';
 import sub from 'level-sublevel';
 import leveljs from 'level-js';
